% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{changeGitHubRepoURLtoGitHubRepoAPICall}
\alias{changeGitHubRepoURLtoGitHubRepoAPICall}
\title{Change the url of a repository to API url of the repository.}
\usage{
changeGitHubRepoURLtoGitHubRepoAPICall(url)
}
\arguments{
\item{repository}{url}
}
\value{
repository's API url
}
\description{
For example, when running this function with https://github.com/octocat/Hello-World as input would return,
https://api.github.com/repos/octocat/Hello-World
}

