% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Commits.R
\name{cleanUpCommitData}
\alias{cleanUpCommitData}
\title{Clean up the commit data provided by getCommitData function}
\usage{
cleanUpCommitData(commitData)
}
\arguments{
\item{object}{created by getCommitData function}
}
\value{
commit.cha, commit.message, commit.date, author.name, author.email, author.login, committer.name, committer.email, committer.login
}
\description{
The getCommitData function returns data in nasty nested list, so cleaning up is necessary
This function gets rid of unnecessary information and returns commit sha, commit message, commit date
repository author's name, email and username and commiter's name,email and username.
}

