% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Issues.R
\name{cleanUpIssueData}
\alias{cleanUpIssueData}
\title{Clean up the issue data provided by getIssueData function}
\usage{
cleanUpIssueData(issueData)
}
\arguments{
\item{object}{created by getCommitData function}
}
\value{
issue title, issue author, issue state, issue assignee, issue label, issue milestone, issue numComments, issue created date, issue body
}
\description{
The getIssueData function returns data in nasty nested list, so cleaning up is necessary
This function gets rid of unnecessary information and returns issue title, issue author,
issue label, issue state, issue assignee, issue milestone, issue comments, issue created date, issue body
}

