% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{divideBy5000}
\alias{divideBy5000}
\title{Divide the elments by 5000 accomodating GitHub's 5000 API call per hour policy}
\usage{
divideBy5000(x, token)
}
\arguments{
\item{x}{: vector of elements, token}
}
\value{
matrix which tells starting and ending position
}
\description{
Divide the elments by 5000 accomodating GitHub's 5000 API call per hour policy
}

