% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{extract_info}
\alias{extract_info}
\title{Using a GitHub oauth2.0 token, execute an API call}
\usage{
extract_info(apiCall, token)
}
\arguments{
\item{apiCall, }{token}
}
\value{
API call output
}
\description{
Using a GitHub oauth2.0 token, execute an API call
}

