% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Commits.R
\name{getCommitData}
\alias{getCommitData}
\title{Get the data of commits for the repository}
\usage{
getCommitData(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
nested list of all commits for the repository
}
\description{
For a given repository url, return data of commits of the repository
}

