% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Forks.R
\name{getForkData}
\alias{getForkData}
\title{Get data of users who forked certain repository}
\usage{
getForkData(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
usernames of those who forked the repository
}
\description{
For a given repository url, return n x 69 dataframe of n users who forked the repositry
}

