% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Issues.R
\name{getIssueData}
\alias{getIssueData}
\title{Get the data of issues for the repository}
\usage{
getIssueData(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
nested list of all issues for the repository
}
\description{
For a given repository url, return data of issues for the repository
}

