% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{getNumForks}
\alias{getNumForks}
\title{Get the number of forks for a repository.}
\usage{
getNumForks(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
The number of forks for a particular repository
}
\description{
For a given repository url, return the number of forks for the respository
}

