% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{getNumStargazers}
\alias{getNumStargazers}
\title{Get the number of stargazers for a repository.}
\usage{
getNumStargazers(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
The number of stargazers for a particular repository
}
\description{
For a given repository url, return the number of stargazers for a respository
}

