% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Orgs.R
\name{getOrgsData}
\alias{getOrgsData}
\title{Get API data of an organization}
\usage{
getOrgsData(orgs.url, token)
}
\arguments{
\item{orgs.url, }{token}
}
\value{
data regarding the given organiztion url
}
\description{
Return data regarding the given organization url
}

