% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Orgs.R
\name{getOrgsTop10Repos}
\alias{getOrgsTop10Repos}
\title{Return urls of 10 repositories with the most stars for the given organization}
\usage{
getOrgsTop10Repos(orgs.url, token)
}
\arguments{
\item{orgs.url, }{token}
}
\value{
ten repositories urls with the most stars
}
\description{
Extracts number of stars for all repositores belonging to the organization then return the top 10
}

