% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Stargazers.R
\name{getStargazersData}
\alias{getStargazersData}
\title{Get data of users who starred certain repository}
\usage{
getStargazersData(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
usernames of those who starred the repository
}
\description{
For a given repository url, return n x 17 dataframe of n users who starred the repositry
}

