% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{getStargazersUsername}
\alias{getStargazersUsername}
\title{Get the names of users who starred certain repository}
\usage{
getStargazersUsername(repository.url, token)
}
\arguments{
\item{repository.url, }{token}
}
\value{
usernames of those who starred the repository
}
\description{
For a given repository url, return the usernames of those who starred the repositry
}

