% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Users.R
\name{getUserdata}
\alias{getUserdata}
\title{Get userdata for the given username(s)}
\usage{
getUserdata(username, token)
}
\value{
number of calls left
}
\description{
Get userdata for a given username(s)
}

