% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{gitHub_Token_Generator}
\alias{gitHub_Token_Generator}
\title{Create a GitHub oauth2.0 token}
\usage{
gitHub_Token_Generator(clientID, clientSecret)
}
\arguments{
\item{clientID, }{clientSecret}
}
\value{
GitHub oauth2.0 Token
}
\description{
This function takes clientID and clientSecret as input and create a GitHub token.
}

