% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{LimitConsciousExtraction}
\alias{LimitConsciousExtraction}
\title{data extraction using GitHub API conscious of the API call rate limit
due to GitHub's 5000 call per hour rate limit, data extraction that requires more than 5000 call can be cumbersome.
This function executes function while automatically accommodating the rate limit.}
\usage{
LimitConsciousExtraction(FUN, vectors, token)
}
\arguments{
\item{token, }{the name of function in characters, x}
}
\value{
list
}
\description{
data extraction using GitHub API conscious of the API call rate limit
due to GitHub's 5000 call per hour rate limit, data extraction that requires more than 5000 call can be cumbersome.
This function executes function while automatically accommodating the rate limit.
}

