% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Basics.R
\name{sleep}
\alias{sleep}
\title{Put system into sleep until GitHub API rate limit has been refreshed}
\usage{
sleep(token)
}
\value{
Puts system in to sleep
}
\description{
Put system into sleep until GitHub API rate limit has been refreshed.
}

