% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GitHub_API_Orgs.R
\name{getOrgsRepoUrls}
\alias{getOrgsRepoUrls}
\title{Return urls of repositories under the specified organization}
\usage{
getOrgsRepoUrls(orgs.url, token)
}
\arguments{
\item{orgs.url, }{token}
}
\value{
URLs of all repositories of the organization
}
\description{
All repositories of the
}

